updating latest version
updating to the latest version of kukulkan

+++ src/main/java/mx/infotec/dads/kukulkan/config/DatabaseConfiguration.java
@@ -49,26 +49,28 @@ public class DatabaseConfiguration {
 	CommandLineRunner init(final DataStoreRepository dsRepository, final DataStoreTypeRepository dsTypeRepository,
 			final RuleRepository ruleRepository, final RuleTypeRepository ruleTypeRepository) {
 		return commandLineRunner -> {
-			log.info("DELETE DATABASES");
-			dsRepository.deleteAll();
-			dsTypeRepository.deleteAll();
-			ruleRepository.deleteAll();
-			ruleTypeRepository.deleteAll();
-			DataStoreType dst = EntitiesFactory.createDefaultDataStoreType();
-			dst = dsTypeRepository.save(dst);
-			log.info("dataStore1:"+dst.getId());
-			DataStore dsValuApp = EntitiesFactory.createDefaultDataStore(dst);
-			dsRepository.save(dsValuApp);
-			log.info("dataStore2:",dsValuApp.getId());
-			DataStore testDs = EntitiesFactory.createTestDataStore(dst);
-			dsRepository.save(testDs);
-			log.info("dataStore3:",testDs.getId());
-			DataStore dsConacyt = EntitiesFactory.createConacytDataStore(dst);
-			dsRepository.save(dsConacyt);
-			RuleType singularRuleType = ruleTypeRepository.save(EntitiesFactory.createDefaultSingularRuleType());
-			ruleTypeRepository.save(EntitiesFactory.createDefaultPluralRuleType());
-			ruleRepository.save(EntitiesFactory.createOsRule(singularRuleType));
-			ruleRepository.save(EntitiesFactory.createEsRule(singularRuleType));
+			if (dsRepository.count() == 0) {
+				log.info("DELETE DATABASES");
+				dsRepository.deleteAll();
+				dsTypeRepository.deleteAll();
+				ruleRepository.deleteAll();
+				ruleTypeRepository.deleteAll();
+				DataStoreType dst = EntitiesFactory.createDefaultDataStoreType();
+				dst = dsTypeRepository.save(dst);
+				log.info("dataStore1:" + dst.getId());
+				DataStore dsValuApp = EntitiesFactory.createDefaultDataStore(dst);
+				dsRepository.save(dsValuApp);
+				log.info("dataStore2:", dsValuApp.getId());
+				DataStore testDs = EntitiesFactory.createTestDataStore(dst);
+				dsRepository.save(testDs);
+				log.info("dataStore3:", testDs.getId());
+				DataStore dsConacyt = EntitiesFactory.createConacytDataStore(dst);
+				dsRepository.save(dsConacyt);
+				RuleType singularRuleType = ruleTypeRepository.save(EntitiesFactory.createDefaultSingularRuleType());
+				ruleTypeRepository.save(EntitiesFactory.createDefaultPluralRuleType());
+				ruleRepository.save(EntitiesFactory.createOsRule(singularRuleType));
+				ruleRepository.save(EntitiesFactory.createEsRule(singularRuleType));
+			}
 		};
 	}
